Refactor extract LoginOfflineLink to 'gui.auth.use_offline' namespace
